Set MAAPI FLAG_DELAYED_WHEN flag
maagic_copy should "just work". In case the destination subtree uses
"when" conditions to hide leaves, we set the FLAG_DELAYED_WHEN flag.
This MAAPI flag postpones the evaluation of when statements until the
moment the write transaction is applied. To preserve the state of the
flags in the transaction(s) backing the user-provided source and
destination maagic objects, we start nested transactions
(trans-in-trans) for the necessary flags.